BOAC-2822, drop-in advisor homepage shares DropInWaitlist component with /appt/desk
